pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linux-commandcrafter"
version = "2.0.0"
description = "Visually build, analyze, and manage Linux commands. Server-rendered Flask rewrite."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "flask>=3.0",
    "gunicorn>=22.0",
    "google-genai>=0.8.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
commandcrafter = "commandcrafter.__main__:main"

[tool.setuptools.packages.find]
include = ["commandcrafter*"]

[tool.setuptools.package-data]
commandcrafter = ["templates/**/*.html", "templates/*.html"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
